Add Soap Auth Client and fix issues with dates and composite collection requests and responses
