+++ Tests/Rest/ClientTest.php
@@ -7,7 +7,7 @@
  */
 namespace AE\SalesforceRestSdk\Tests\Rest;
 
-use AE\SalesforceRestSdk\AuthProvider\LoginProvider;
+use AE\SalesforceRestSdk\AuthProvider\OAuthProvider;
 use AE\SalesforceRestSdk\Rest\Client;
 use PHPUnit\Framework\TestCase;
 
@@ -21,7 +21,7 @@ class ClientTest extends TestCase
     protected function setUp()/* The :void return type declaration that should be here would cause a BC issue */
     {
         $this->client = new Client(
-            new LoginProvider(
+            new OAuthProvider(
                 getenv("SF_CLIENT_ID"),
                 getenv("SF_CLIENT_SECRET"),
                 getenv("SF_USER"),
@@ -46,7 +46,7 @@ class ClientTest extends TestCase
 
         $this->assertNotNull($limits);
 
-        $class = new \ReflectionClass(LoginProvider::class);
+        $class = new \ReflectionClass(OAuthProvider::class);
 
         $tokenProperty = $class->getProperty('token');
         $tokenProperty->setAccessible(true);
