+++ src/Model/Rest/Composite/SObject/GetUpdatedSubRequest.php
@@ -113,9 +113,12 @@ class GetUpdatedSubRequest extends GetSubRequest implements GetUpdatedSubRequest
         $this->url = '/'.Client::BASE_PATH.'sobjects/'.$this->sObjectType.'/updated/?'
             .http_build_query(
                 [
-                    'start' => $this->start->format(\DATE_ISO8601),
-                    'end'   => $this->end->format(\DATE_ISO8601),
-                ]
+                    'start' => $this->start->format('Y-m-d\Th:i:sP'),
+                    'end'   => $this->end->format('Y-m-d\Th:i:sP'),
+                ],
+                null,
+                '&',
+                PHP_QUERY_RFC3986
             );
     }
 }
