+++ src/Rest/Client.php
@@ -26,7 +26,7 @@ use Psr\Http\Message\RequestInterface;
 
 class Client extends AbstractClient
 {
-    public const VERSION = "43.0";
+    public const VERSION = "44.0";
 
     /**
      * @var string
